# Conflicts:
#	scripts/stonedb_build.sh
#	sql/log_event_push_cond.cc
#	sql/sql_load.cc
#	storage/tianmu/core/engine.cpp
#	storage/tianmu/core/transaction.cpp

+++ storage/tianmu/core/group_distinct_table.cpp
@@ -57,8 +57,8 @@ void GroupDistinctTable::InitializeVC(int64_t max_no_groups, vcolumn::VirtualCol
   DEBUG_ASSERT(!initialized);
   if (max_bytes > 0)
     max_total_size = max_bytes;
-  if (max_bytes > 2_GB)  // possible for large aggregation settings, but
-                         // not allowed here - limit to 1 GB
+  if (max_bytes > static_cast<int64_t>(2_GB))  // possible for large aggregation settings, but
+                                               // not allowed here - limit to 1 GB
     max_total_size = 1_GB;
   if (max_no_rows == common::NULL_VALUE_64)
     max_no_rows = 0;  // not known
